Add e.stack to test failure messages where available
This makes the failure message unreasonably verbose (in Gecko and WebKit, the
only place it works).  But it doesn't seem to happen for DOMExceptions, so
generally if you get the message it's because your test has a bug in it.

Lightly tested in the forked dom-range copy of the test harness.  If there are
any problems, please revert and tell me.