Use a CSS rule to hide cells, not zillions of style attributes
This vastly speeds up the performance of clicking the show/hide checkboxes in
all browsers.  In everyone but Firefox, it's just a second or two even for a
multi-thousand-line table.  It doesn't seem to help Firefox too much, though.

Lightly tested in dom-range like all the others, revert if there are problems.