Rewrite assert_throws()
As discussed in #whatwg:

http://krijnhoetmer.nl/irc-logs/whatwg/20110427#l-960

James doesn't like it, but I convinced him to let me commit it for now so that
I can make DOM Range tests actually work in the short term.  At some point he
wants to rewrite it.

Features of the new version: clearer error messages, and supports magic
pass-just-a-string syntax for RangeException.