Revert 18d2f1031deb, run many tests per attribute
This means that in IE and Opera, you don't just have every reflection test fail
because of one bug in all of them (like, say, mishandling nulls in strings).
Requested by jgraham and others.  This makes the testharness.js-based version
work more like reflection-original.js, although it's still less useful when
used standalone because it doesn't aggregate similar errors.  (But it's more
useful if you want to do machine processing on the test results, or are just
used to testharness.js.)