Define indeterminate for insert*List
Now everything that has either a state or value defined also has
indeterminate defined, except styleWithCss (where it doesn't make
sense).  The meaning as a rule is that it's indeterminate if either the
state is true for part of the selection but not the whole selection, or
there are multiple values in the selection, as the case may be.