Make "remove preserving descendants" block-only
It was only used in one place in the inline commands, and the
implementation was never actually updated to use it.  All the fiddling
it does is only useful for block formatting anyway.  This is mostly just
moving code around, but it paves the way for a substantive change:
preserving styles when I split the parent.