Preserve styles when breaking up nodes
This is a pretty big change.  It mostly only does anything when there
are inline style attributes.  So now when you have
  <p style=color:red>foo<p>[]bar
and hit backspace, you get
  <p><font color=red>foo</font>[]bar
or
  <p><span style=color:red>foo</span>[]bar
instead of
  <p style=color:red>foo[]bar
which makes "bar" red too.  In principle it might have effects in other
scenarios too.  If it doesn't, it might be best to just aggressively
kill inline style attributes on blocks . . .