Commit a whole bunch more stuff
The theoretical reason for this commit was to support state and value
for createLink and unlink.  Other stuff that happened along the way was
improving valuesEqual, fixing a formatBlock bug while also making it
more readable, making my list of invariants more correct, and making the
thing that outputs the indeterm/state/value work a lot better.

I wanted to use git add -i to break this up into a bunch of unrelated
commits, because this kind of big messy commit gets on my nerves, but
nobody's likely to be reviewing these in much detail, so it's really not
worth the time . . .