Rewrite getDescendants() non-recursively
Chrome keeps freezing on insertOrderedList and insertUnorderedList
tests.  When I tried pausing in the debugger, it actually worked for
once, and displayed a stack like a dozen or more deep of
getDescendants(), then it crashed the tab.  Making it non-recursive
seems to improve things.  (Specifically, it was calling
getDescendants(document), which is a worst case and I should probably
get rid of it.)