Add script to automatically fix up data.js
This removes a tedious manual step from regenerating the tests.  Now my
procedure is:

1) Run gentest.html in Chrome and Firefox.  This will take 15ish minutes
on this machine (sigh), with Chrome being the limiting factor because
it's well-behaved and nices the background tab.  I notice when it's done
because of the CPU meter I have at the top of my screen.

2) Switch to some other branch on my server so I can keep working.

3) When (1) is done, copy results to /tmp/webkit and /tmp/gecko on
localhost.  I do this locally because trying to paste 1MB of text into a
remote terminal seems not to work well.

4) diff -u /tmp/webkit /tmp/gecko | less on localhost.  Check for any
unexpected differences.  There are some known mismatches due to browser
bugs, which are mentioned in the spec.  (I should automate this step
too, perhaps?  It could be error-prone.)

5) Switch back to the branch I was working on on the server.

6) scp /tmp/webkit to the right place on my server, overwriting data.js.
I use WebKit instead of Gecko because it has fewer bogus results that
I've noticed.

7) conformancetest/fixdata on the server to fix it.

8) git diff to check that the test changes all look correct.

9) git commit -a --amend to add the test changes to the existing commit.

10) Rebase the branch onto master, delete the branch, and ./publish.

I'm going to have to think of more ways to simplify this process,
although with branches it's surprisingly painless.