Don't wrap invisible nodes when forcing style
In the course of doing this, I had to also make state and indeterm
ignore invisible nodes.  Otherwise, you'd bold something like
  <p>[foo</p> <p>bar]</p>
and get
  <p><b>[foo</b></p> <p><b>bar]</b></p>
which would have indeterm true and state false, because of the editable
text node " " in between the paragraphs.  I introduced a new concept of
"formattable node" to make this cleaner and allow easier future changes.

I also rationalized the boilerplate tests at the beginning of each
inline formatting command, making them all the same.

One lingering issue I don't like so much: the test
  <p>[foo</p><p> <span>bar</span> </p><p>baz]</p>
becomes something like
  <p><b>[foo</b></p><p> <b><span>bar</span> </b></p><p><b>baz]</b></p>
with the first invisible node outside the <b> but the second inside.
This is because we go through the nodes sequentially, and do merge
invisible nodes into adjacent wrappers, but don't create a wrapper for
them.  So we get something like
     <p> <span>bar</span> </p>
  -> <p> <b><span>bar</span></b> </p>
  -> <p> <b><span>bar</span> </b></p>
where the first " " is unaffected because there's no adjacent <b> yet,
but the second gets stuck into the <b> added for the span.  This isn't a
big deal, but it's a bit weird, and worth revisiting.  I haven't marked
the relevant tests as bad, but perhaps I should.

Fixes: http://www.w3.org/Bugs/Public/show_bug.cgi?id=13996