Partial update of reflection tests
Gave up midway through because it's just too much work.  There are probably
still a bunch of things that are outdated or otherwise erroneous.  The
requirements in the spec are too scattered to deal with easily: you have to
jump all over the place to figure out what types attributes have, and things
like that.