Fix testharness.js in IE/Opera
They broke because IE doesn't support setting innerHTML at all on <table> or
<tbody> (?!), and Opera was producing an incorrect DOM due to use of a
non-HTML5 parser (<td><td> would nest instead of auto-closing the first <td>).
A side effect of this change is that the header row of the results table is put
in a <thead>, when formerly the <tr> was directly a child of the table, but
that shouldn't hurt anything.