Make format_value faster and less verbose
This makes output a lot shorter for some nodes, and across the board for
Ranges/Selections.  It removes the special case for Range/Selection that I just
added, because it made things very verbose, and was fairly useless given the
way I was writing my tests.

Additionally, it cuts Chrome's execution time on my test case from 5s to 3s, in
the case where there's no <div id=log> (as might be in browser regression
tests).  format_value previously took up 50% of total execution time and now is
10%.  25% of time is now spent in addRange, which is all the "real" work that's
being done, so that's good.  A quarter of the time is still spent in
make_message, but about half of that is format_value, so I don't see huge
potential gains in Chrome by optimizing the rest.  So I consider Chrome
optimization for this particular test case basically done, other than layout
time.

Now I'll look at Firefox, which shows different slowdowns.