Merge adjacent lists when deleting
This does not match current browser behavior, but makes more sense from
a user perspective.  Note that we merge the lists even if they were
adjacent to start with, and we didn't make them so by deleting anything.
Adjacent lists of the same type don't make sense for us -- it shouldn't
be possible to create them using insertOrderedList/insertUnorderedList,
so it shouldn't be possible to create them using delete either.

Fixes: http://www.w3.org/Bugs/Public/show_bug.cgi?id=13976