Completely revamp idlharness.js
This adds lots more features, and lays the groundwork for many more features to
come.  The API changes make it completely incompatible with yesterday's
version; more work remains to be done.  I will push a rewrite of
http://w3c-test.org/webapps/DOMCore/tests/submissions/AryehGregor/interfaces.html
shortly that updates it to use the new version.  Comments on the API are
greatly appreciated, as are feature requests and such.  I intend it to be
usable for the HTML spec tomorrow, if all goes well -- currently it's more or
less useless except for DOM4.