More avoiding of uninformative exceptions
Previously, assert_inherits(location, "href") would throw a somewhat
inscrutable exception in Gecko.  I tracked it down to the fact that
window.location is an object and is instanceof Object, but
window.location.hasOwnProperty is undefined (?!).  An extra sanity check makes
the error much more comprehensible.

Yes, this is a very weird and isolated bug to add a sanity check for.  But I
think it's worthwhile if it comes up in a real-world test suite.