Update insertNode tests for spec change
https://www.w3.org/Bugs/Public/show_bug.cgi?id=15350

Previously, none of the test ranges started in a comment, so this line in the
spec wasn't actually tested (oops).  This commit adds a new test range, for
which most of the tests would have failed in all browsers.  But it also updates
the insertNode test implementation, after which Firefox 12.0a1 and Chrome 17
dev (and probably IE too, didn't test) pass all the newly-added tests.  AFAICT,
no preexisting tests newly fail, so the spec change is a strict compat
improvement (although I didn't examine too closely).