Various minor idlharness tweaks
Make repeated add_objects calls actually work.  Don't try to do anything magic
with constructors; we can't guarantee from the IDL that a constructor call
won't throw, so it's more useful for the library user to explicitly provide
constructor calls that aren't supposed to throw.  And test for the absence of
getters/setters in an extra place.