Increase precision for computed style
Previously matrix(0, 0, 0, 0, 0, 0) was equivalent to matrix(1, 0, 0, 1, 0, 0)
because of the tolerance of 1.5.  I've left that tolerance for pixel
measurements, but cranked down the tolerance for computed values to 0.05.  This
is enough that one additional test fails in Gecko due to rounding, a bunch more
fail in Opera, none in IE, and (I think) none in WebKit.  (A couple hundred
tests newly fail in WebKit because they were actually 1's instead of 0's or
similar, but I didn't spot any that fail because of too much rounding.)

I also fixed a few bugs in the tests that happened not to be very noticeable
because the tolerances were too high.