Test multiple 3D transforms
It's hard to tell if these tests are all correct, because in some cases both
Gecko and WebKit fail.  But they seem to fail in entirely different ways in the
cases I looked at where they both failed, so my confidence in the tests isn't
entirely undermined.  There are just lots of bugs here.