Start porting Mozilla's transition tests
I was working on this for a while now, actually.  This is just a commit so that
my work is saved.  The tests currently don't run with stock testharness.js; you
need the assert_approx_equals fix from <https://github.com/ayg/testharness.js>,
hopefully soon to be pulled.  There's still quite a bit of porting work to be
done.  I'm rewriting the file both for readability (keeping each test type
self-contained), and to remove Gecko dependencies.  And, I admit, converting it
to my preferred coding style as I go, out of habit.