shadow-dom: Normalize the line endings to LF.
This change normalizes the line endings of the files under shadow-dom/
directory to LF, which I believe is standard configuration of Git.

In addition, this change adds the .gitattributes file which lets Git deal with
newlines automatically.

See also: <https://help.github.com/articles/dealing-with-line-endings>

No change is made to the content of tests.