shadow-dom: Add test conversion script.
This script converts tests under the shadow-dom/ directory to separate HTML
files with appropriate metadata format. The output is produced within
the out/ directory by default to avoid conflicts with existing files, but
I intend that the files under the out/ directory will take over the existing
contents of the shadow-dom/ directory.

The script file will be removed after the conversion.