shadow-dom: Reorganize the test structure.
This change reorganizes the tests under the shadow-dom/ directory. In the new
format, each test has a dedicated, clean HTML with appropriate metadata.
The new format provides a clear directory structure based on the section
hierarchy of the specification. In addition, numerous issues with the existing
tests are addressed, such as unnecessary files, messy metadata definitions,
and so on.

The new files are generated by convert.py, which is included in the previous
commit.

There are some issues I could not address using a script; most notably
each test does not have a sensible name. These issues need to be addressed
manually, so I left these as future work.