shadow-dom: Handle vendor prefixes in a more sophisticated way.
This change adds a functionality to testcommon.js that enables the tests to
handle vendor prefixes without manual intervention. When testcommon.js is
loaded, it automatically adds unprefixed methods and attributes if necessary.

This change obsoletes the existing utility functions related to vendor
prefixes, but they cannot be removed at this moment, because there are many
tests using them. These functions will be removed eventually.