I adjusted c541-word-sp-000.xht, c542-letter-sp-000.xht and c542-letter-sp-001.xht to be using a computed font-size which would be dividable by 5 and by 2.
I created reftests c541-word-sp-000-ref.xht
and c542-letter-sp-000-ref.xht .

Note that Opera 12.15 does not handle well the
value 22.5pt in c541-word-sp-000.xht and in
c542-letter-sp-000.xht.

What remains to do now is to add the "may" flag
in c541-word-sp-001.xht and in c542-letter-sp-001.xht
because of the use of negative value in those tests.