shadow-dom: Improve the readability of two basic reftests.
This change updates two basic reftests in the shadow-dom directory to make
them easier to understand. The new tests are intuitive (green is pass and
red is failure), and has a sentence explaining the passing condition.

This change also fixes some style issues. Most notable error was the name of
link type for reference ("reference" is wrong; "match" should have been used).