shadow-dom: Refine script tests related to ownerDocument property.
This change updates the existing test that checks the ownerDocument property.
The test was split into two files, and renamed to ownerdocument-###.html to
better reflect the test content. Also, the test names were updated so people
can understand what each test is doing.

This change is basically a rewrite of the existing tests, but some tests
were enhanced in hope of achieving better coverage.

ownerdocument-001.html contains basic test cases for ownerDocument property.
ownerdocument-002.html is a test that checks whether the ownerDocument property
is working properly on every HTML5 element.