shadow-dom: Remove a test file for Selectors API.
This change removes test-008.html. The tests in this file are (mostly) covered
in selectors-api-001.html.

test-008.html includes tests that pass complex selectors to querySelector()
and querySelectorAll(), but it is suspected that these tests add little value;
we are not testing Selectors API itself but testing interactions with it, thus
the existing tests in selectors-api-001.html should suffice for this purpose.