shadow-dom: Add reftests for distribution on DOM changes.
Shadow DOM distribution should re-occur when any variable that affects the
distribution has changed. This change adds a series of reftests that test
this condition.

It is infeasible to cover every possible case where distribution must re-occur,
but tests added in this change should cover most casual cases.

The old test test-004.html is removed in favor of the new tests.