shadow-dom: Provide proper arguments to DOMImplementation.createDocument().
Existing code did not give an argument to DOMImplementation.createDocument().
However, this is incorrect according to the definition of createDocument(),
which requires at least two arguments.

It seems that Chromium has recently tightened up a certain kind of invocations
to DOM APIs. Some tests started to fail on Chromium for this reason.