Order is apparently wrong making the test fail.
Based on my understanding of the tabindex behavior as explained in the
html5 living standard: pre-order, depth first tree order ("before any
element whose tabindex attribute has a value equal to the value of the
tabindex attribute on the element but that is later in the document in
tree order than the element"). Confirmed that other browsers (FF, IE,
Safari) follow the same tabindex order on the final document object.