Add a test for an edge case resource selection invocation in Blink
In current Blink, inserting a media element with a src attribute and
networkState NETWORK_EMPTY will trigger resource selection. Normally the
src attribute and NETWORK_EMPTY don't go together, but when a media
element is moved between documents its networkState is reset. None of
this is per spec, of course.