Test for abort and emptied events in HTMLMediaElement.load()
Thanks to Revlin John who submitted a similar test:
https://github.com/w3c/web-platform-tests/pull/571

This started out as improvments to that test, but quickly morphed beyond
recognition in an attempt to provide full coverage.

There's plenty of load() left to test, but that is better left to
dedicated tests since they will require more complex setup to e.g. get
in-band text tracks or seek in the video.