Add tests for SVG sizing
Sizing of inline SVG and SVGs inside <object>, <iframe> and <img>
lacks interoperability between browsers. HTML, CSS 2.1 and SVG all
involved to some degree here, but the basic principles comes from CSS
2.1, Visual formatting model details.

The tests generates a couple of scenarios (each a TestData) for sizing
SVG inside <object>, <iframe>, <img> and inline in the html. A simple
JavaScript sizing implementation is used to calculate the expected
size. The generated DOM is probed to get the actual size.

To debug a specific test append ?id=<test-id> to the URLs.