Always close the connection at the end of a request.
Since changing the HTTP version to 1.1, the connection would use
keepalive. This is fine in most cases, but some responses don't
know how long they are and so don't set Content-Length. This meant
that the associated request would time out. Closing the connection
fixes this issue. If we need keepalive e.g. to test browser behaviour
in this area then we can revisit this fix with something more
surgical.