assert 4th argument is a non-specific number
N'either the ECMAScript specification nor the web-apps/workers specifications specify specific rules about what the column number actually points to.

Maybe they should, but they don't. Column 5 (the position of the "throw" keyword) is not necessarily more helpful than the column of the expression being thrown, if any. In any case, since it's not specified anywhere, an exact number should not be asserted. What should be asserted is that the property exists at all