Add support for configuring supported browsers through ini file.
This adds much more comprehensive support for ini files, and the ability to
override the ini file through command line arguments. It also adds the
ability to specify which products are supported through the [products]
section of the ini file. This provides a way of not requiring dependencies
for those products that won't be used in a given environment.