Ensure that just reading a manifest file doesn't require us to be in a git repo.
This regressed with ato's recent changes. In general it isn't OK to assume that
web-platform-tests are actually running in a git repo rather than in a copy
except when doing operations that actually require a git operation. Therefore
trying to find the repo root at the top level isn't OK.