Server changes for to allow mounting directories at a non-root path.
This allows routes like /foo/ that correspond to some filesystem path
outside the doc root. To make this work, there is a
non-backward-compatible change to the way that * is handled in path
expressions; it now matches the whole rest of the path in all cases, so
/foo/*.py matching /foo/bar/baz.py would have a match group of *
containing bar/baz.py rather than bar/baz.