Added --config option to serve.py.
In addition to reading the standard config.default.json file and any
config.json file in the working directory, this option allows you to
specify a custom location for any additional config file to read.
Meant to be useful for the case of, e.g., you don't want to have to keep
settings in a custom config.json in your working directory due to risk of
it getting blown away by a "git clean" or whatever.