cc Dropped two tests that aren't actually useful.
These two tests were meant to be testing bad surrogate pairs in values of
the accesskey attribute. But the files are encoded in UTF-8, so no chars in
the surrogate-pair ranges are conforming in them anyway. So these tests are
really only parser tests, because they'll just cause a conforming parser to
emit a "bad byte sequence" (or whatever) message long before any checker
can do other specific checks on the value of the accesskey attribute.