Fix flex-margin-no-collapse.html
As described in https://lists.w3.org/Archives/Public/www-style/2015Mar/0463.html, the testcase is broken. This patch adds flex:none so that it displays correctly; this seems to match the intention better than the alternative of changing the container's height to 400px. Without this change, the boxes will shrink due to the flexing and not cover up the red box.