Add EventWatcher helper to testharness.js
It is very difficult to test that a series of events occurs in a very specific
order, while also checking that unexpected events do not occur between any of
the expected events. This helper can create Promises that resolve when the
expected events occur, and reject when unexpected events occur, which makes
the process of testing such events much more feasible.