I modified the float-contiguous-vrl-010.xht and float-contiguous-vlr-011.xht in a way that bypasses the problem of the line-height used value of the pass-fail-conditions sentence by Chrome (19px) and Firefox (20px).
If that workaround is incorrect (due to
latitude granted to browsers for abs. pos.
left: auto and top: auto), then I can
as a last resort wrap that abs. pos.
div#reference-overlapped-red into
a rel. pos. empty div but right now
I do not see the need for that or
why I should do that.