Change implementation of EqualTimeChunker.
The new implementation divides the list of tests into N segments with
equal numbers of tests. It then tries to minimize the badness score,
defined as sum((chunk_time - sum(chunk_time)/N)**2), by moving one subdir
of tests at a time into a neighboring chunk if doing so produces a
decrease in the badness.

This algorithm obviously only finds local minima of the badness.