WebRTC: Data channel test for empty string messages
This test creates a Data channel between two local PeerConnection instances
and checks how empty string messages (a call to "send('')" in other words)
are processed.

The steps to follow to send a message (rightly) do not distinguish between
empty strings and other types of strings, so empty strings sent on one end
should be received on the other end, see:

 http://w3c.github.io/webrtc-pc/#dom-datachannel-send

The test currently fails in Chrome and Firefox.

Note that the behavior described above is consistent with how empty strings
are managed in WebSockets: calling "send('')" on an opened WebSocket correctly
triggers a "message" event on the receiving side.