UI Events first exploratory event-order test
Adds a simple test for mousemove testing using the EventRecorder utility
(also committed as part of this change).

The UIEvents editors are exploring approaches to a test-driven spec
process. This first test will be used by the group to take initial
feedback on a testing approach including what limitations or addition
features may be required.