Fix combination of --include and --exclude.
Manifests have the semantic of either looking at the value from
the current key, or the value from the root (not the value from
the parent). This is necessary for expected keys to work well,
but isn't really what's needed for include manifests. In particular
the --include and --exclude functions were acting as if we would
traverse up the parent node chain of the manifest until we found
an explicitly set skip key. Instead we make sure that skip is
set on every node in the chain.