Removed existing Notification dir（manual）tests.
As-is, these tests aren't testing what they claim to be testing and anyway
the spec doesn't actually have clear normative/testable requirements for
UAs related to text direction in notifications, except just requiring that
UAs use the value of the `dir` option to Notification constructor to set
the `dir` property on Notification instances. And the existing manual
dir-*.html tests this commit is dropping aren't testing that requirement
anyway (we already have other tests that are).

Beyond that, given the spec doesn't clearly define any other requirements
for what else UAs must do with text direction, along with the fact that
whatever happens with the text of displayed notification is
platform-dependent and not testable through just observing the UA behavior,
I don't think we can currently construct any other useful tests for it.