Reorganize for merging to web-platform-tests
There is no usable editing spec, so the editing tests are really
regression tests, not conformance tests, as noted in the README.  The
selection tests are based on the selection spec when I had it, so may be
out-of-date if Ryosuke has changed anything.

A ton of code in editing/run/implementation.js (practically the whole
file) and tests.js is now unused, because it was used for test
generation debugging etc. and is not used for running conformance tests.
I didn't bother deleting it.