Use deepPath for event path trimming.
What this test (test-003.html) does is now specced as
"Event Path Trimming", and when retargeted "target" and
"relatedTarget" become same, not only trusted events are
not dispatched, Event.deepPath is also trimmed for these
nodes.

Now the test uses .deepPath rather than relying on
generating "trusted" events and checking if event handler
is not invoked.