Fix #2280: Update HTMLScriptElement event and htmlFor tests
Update the tests for the event and htmlFor IDL attributes of the
HTMLScriptElement to expect the correct behavior:

The .event and .htmlFor IDL attributes of the script element should reflect
the element's "event" and "for" content attributes (respectively),
and not return the empty string on getting and do nothing on setting
as previously stated.

See whatwg/html#253 for the related correction to the spec.