[css-grid] Fix test for flexible lengths under 1
The spec now says (http://www.w3.org/TR/css-grid-1/#flex-factor-sum):
"Let flex factor sum be the sum of the flex factors of the flexible
tracks. If this value is less than 1, set it to 1 instead."

This means that flexible tracks under 1fr will have an special behavior.
Updated the tests to follow this new behavior.