Simplify various keyframe-effect tests;
This patch just simplifies the keyframe-effect tests so that we don't have to
repeat default values. This makes the tests shorter, easier to scan, and
easier to understand what is being tested.

In some cases we still repeat the default values in order to indicate that
we're testing that we get a particular default value.

Upstreamed from https://bugzilla.mozilla.org/show_bug.cgi?id=1249212