Check point work
Basic testing now works for coverage on these.  Have not yet done
anything about testing works so still need to look at that.
Currently spec has issue which has related but about using values other
than rsaEncryption for import/export of keys but since the spec says do
it, it is part of the tests