Fix the test for parsing a template element inside a XML docuemnt.
The test previously asserted that template.content.ownerDocuemnt is that of the parsed document.

This patch negates the assertion per
https://html.spec.whatwg.org/#template-contents
https://html.spec.whatwg.org/#appropriate-template-contents-owner-document
which always uses the appropriate template contents owner document for template's content fragment.

The new behavior matches that of Firefox (47.0a1 or 44.0.2) and I'm changing WebKit's behavior in webkit.org/b/148850.