template-as-a-descendant.html erroneously asserts that frameset.innerHTML accepts a template element
Fix the test by replacing the test case to assert that the start tag and the end tag tokens of template element are ignored.

The "in frameset" insertion mode doesn't accept either a start tag or an end tag whose tag name is "template" [1].
Since HTML fragment parsing algorithm [2] starts the HTML parser with a start tag token whose tag name is "frameset",
innerHTML on a frameset element parses the content of innerHTML using the "in frameset" insertion mode.
As a result, start tag whose tag name is "template" is a parsing error and should be ignored.

New behavior matches Firefox (47.0a1) and I intend to change WebKit's behavior in webkit.org/b/143519.

[1] https://html.spec.whatwg.org/multipage/syntax.html#parsing-main-inframeset
[2] https://html.spec.whatwg.org/multipage/syntax.html#parsing-html-fragments