Fix two HTML template element tests for "in frameset" mode.
In the latest HTML specification, "in frameset" mode treats any tag with name 'template' as a parsing error:
https://html.spec.whatwg.org/multipage/syntax.html#parsing-main-inframeset

Updated end-tag-frameset.html's assertion so that it expects no template element be generated under frameset,
and removed a test case in template-owner-document.html which were testing the owner document of a template element
under a frameset element as the parsing beahvior is already tested by the other test.

The new behavior matches that of Firefox, and I'm intending to change WebKit's behavior in webkit.org/b/143519.