webdriver: move existing tests to /old-tests
With the integration of a WebDriver specification test harness into
wptrunner, the tests will have to be rewritten to be compatible with
the new pytest runner.

In order to not have the manifest parser pick up the existing tests,
moving them to /old-tests while we do the conversion/triaging makes sense.