Merge DOMEvents and uievents directories.
Since the DOM Level 3 events spec is now the uievents spec it is
confusing to have two separate directories.

Remove redundant tests; and order the current tests into the correct
directories. I need to iterate on the legacy-domevents-tests and
make them consistent with the rest of the web-platform-tests. Currently
they don't follow the assertion behaviour and manual naming of tests.