Add a test for scrolling to a fragment inside a shadow tree.
As discussed on https://github.com/w3c/webcomponents/issues/66 and https://github.com/whatwg/html/issues/1192,
the user agent should not use an element that has an ID exactly equal to decoded fragid or an anchor element
with a name attribute exactly equal to fragid if the element inside a shadow tree as the indicated part of the document.

WebKit nightly builds and Google Chrome Canary both passes the test.