encoding legacy directories (#3064)
The overwhelming majority of the cases are tabs for the somewhat more complex than usual javascript that drives the tests. I wanted to retain those so that i can read the code easily. I've run all the tests sufficiently often already to believe that these tabs aren't affecting the results.