Delay the SourceFile import in lint.
Due to the way this code is used in web-platform-tests, the localpaths
import is required in that context, but not in tests in this repository.

This introduces the unfortunate problem that check_parsed() can only be
called through main(), but this is the fastest approach I can think of that
will allow the tests in this repository to keep working, while also
ensuring the lint script in web-platform-tests works.