Make suppressing event handling on a document also block script execution for that document.
Note that we do this here instead of when we suspend timeouts, because it makes
it simpler to handle situations in which a window's document changes while
things are suspended/blocked.

The nsDocumentViewer.cpp is not strictly needed, but avoids some extraneous
calls to SuppressEventHandling with a 0 suppression count.

Upstreamed from https://bugzilla.mozilla.org/show_bug.cgi?id=1267989