Chromium tests converted to web platform tests
This pull request converts MSE tests from the Chromium test suite. Conversion
was mostly straightforward. I updated a couple of tests though:

- mediasource-appendbuffer-quota-exceeded.html now uses the usual test.mp4 or
test.webm media file.
- mediasource-avtracks.html tests are more atomic. Typically, removal now only
tests removal and thus no longer fails when the kind, label or language of the
tracks are not valid. By the way, the test assumes that the track language will
be "eng", and I'm not clear whether "en", which Edge reports, is really invalid.