Added assertionCache loading and tracability (#3298)
JSON Schema referenced in assertions by name (e.g., "foo.json") were
being loaded but not placed in the cache.  This meant that they were not
actually being used when testing.

Also improved the output when there are subtests to make it easier to
evaluate which assertion was being evaluated (since assertions are
evaluated bottom-up if there is complex nesting).

Finally, also changed relative URI handling so that we know the complete
path of each component loaded to make it easier to debug.