Added ability to create placeholder tests
Tests named "ph-*.test" will generate .html files that run
automatically but don't actually test anything - returning a NOTRUN
result.  The purpose of this is to support "placeholder" test cases
that can be then populated by members of the working group.

Also added support for "automated" test cases - ones that run fully
automatically when content is provided.

Added examples of these as tests for the updates to make_tests and
to JSONtest.js