Add a lint to check that open() and file() calls in Python specify a mode.
Authors mistakenly using open() in text mode has caused a number of bugs
on Windows. It is rather hard to lint for correct usage here since
text-vs-binary is situation dependant. However it is quite possible to
lint that the author made some explicit choice in the subset of cases
where raw open() is called.