Upstreaming a fix made to the WebKit repository for https://webkit.org/b/160689.
When the test samples 25th pixel it expect the value to be exactly 25% interpolation of the color values.
However, the sampling location is the center of the pixel, i.e. it should correspond to 25.5% interpolation,
which would expect the color of (189.975, 189.75, 65.25, 65.25). The new values are a rounded values of
interpolation “by hand” at steps of 25.5%, 50.5% and 75.5%.

Original fix by Bartosz Ciechanowski.