dom/ranges/Range-mutations.html should not rely on a specific Selection API behavior
This patch updates dom/ranges/Range-mutations.html to compute the expected
result after the DOM mutation based on the actual selection range. In Firefox,
the input range and the selected range are identical. However, the WebKit and
Blink implementations of the Selection API differ and the actual selection
range may end up being different.