Correctly test initEvent unsetting propagation flags (#3472); r=Ms2ger
There was previously never any test for unsetting the immediate
propagation flag.  There was one for the propagation flag, but it called
dispatchEvent() in between initEvent() and the actual test, and since
whatwg/dom#219, dispatchEvent() clears these flags as well, so the test
is now incorrect.  (Before that the test was also not very good, because
Gecko always cleared the flags in dispatchEvent() but not initEvent(),
so it was incorrectly passing the test.)