cc Rename some SVG *-novalid files to *-valid
The files being renamed here all check as valid SVG-in-HTML with the
HTML checker, and none of them otherwise contain any indication of what’s
supposedly invalid about them, so there’s no way for me to tell if the
checker is missing some check it should be doing.

Regardless, the Python build file that was used to generate these tests
indicates that invalidity was evaluated against the SVG DTD. But the
checker using the SVG RelaxNG schemas, not the DTD. So I dunno maybe there
are cases that the DTD prohibits but the RelaxNG schema allows.

Anyway, if the expected errors for any of file thought to be invalid here
can be documented, then we can rename that file back to *-novalid, and I
can add a corresponding check to the validator for that case.