Do not assume that currentTime is still at seeking position when seeking event is fired.
The seeking attribute is changed synchronously, however, the seeking event is fired asynchronously. It is possible that by the time the seeking event is fired, the seeking operation has already completed and that the "time marches on", causing currentTime to be past the seeking position.

MozReview-Commit-ID: IB83w6zKCLF

Upstreamed from https://bugzilla.mozilla.org/show_bug.cgi?id=1293186